--- src/main/java/site/onandoff/member/dto/DBUsing.java
@@ -1,4 +1,0 @@
-package site.onandoff.member.dto;
-
-public interface DBUsing {
-}
