+++ src/main/java/site/onandoff/member/dto/SignUpForm.java
@@ -5,6 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import site.onandoff.member.validator.DBUsing;
+import site.onandoff.member.validator.EmailDuplicateCheck;
+import site.onandoff.member.validator.EmailForm;
+import site.onandoff.member.validator.EmailSize;
+import site.onandoff.member.validator.NicknameDuplicateCheck;
+import site.onandoff.member.validator.NicknameForm;
+import site.onandoff.member.validator.PasswordForm;
 
 @GroupSequence({SignUpForm.class, DBUsing.class})
 @NoArgsConstructor
