+++ src/main/java/site/onandoff/member/validator/DBUsing.java
@@ -1,0 +1,4 @@
+package site.onandoff.member.validator;
+
+public interface DBUsing {
+}
