+++ src/main/java/site/onandoff/member/validator/EmailDuplicateCheck.java
@@ -1,4 +1,4 @@
-package site.onandoff.member.dto;
+package site.onandoff.member.validator;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -14,32 +14,37 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import jakarta.validation.Payload;
 import site.onandoff.member.infrastructure.MemberRepository;
+import site.onandoff.util.encryption.AES256Manager;
 
 @Documented
-@Constraint(validatedBy = NicknameDuplicateCheck.NicknameDuplicateValidator.class)
+@Constraint(validatedBy = EmailDuplicateCheck.EmailDuplicateValidator.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
-public @interface NicknameDuplicateCheck {
+public @interface EmailDuplicateCheck {
 
-	String message() default "이미 존재하는 닉네임입니다.";
+	String message() default "이미 존재하는 이메일입니다.";
 
 	Class<?>[] groups() default {};
 
 	Class<? extends Payload>[] payload() default {};
 
 	@Component
-	class NicknameDuplicateValidator implements ConstraintValidator<NicknameDuplicateCheck, String> {
+	class EmailDuplicateValidator implements ConstraintValidator<EmailDuplicateCheck, String> {
 		private MemberRepository memberRepository;
+		private AES256Manager aes256Manager;
 
 		@Autowired
-		public void setMemberRepository(MemberRepository memberRepository) {
+		public void setMemberRepository(MemberRepository memberRepository, AES256Manager aes256Manager) {
 			this.memberRepository = memberRepository;
+			this.aes256Manager = aes256Manager;
 		}
 
 		@Override
-		public boolean isValid(String nicknameInput, ConstraintValidatorContext context) {
-			return !memberRepository.existsByNickname(nicknameInput);
+		public boolean isValid(String emailInput, ConstraintValidatorContext context) {
+			String encryptedEmailInput = aes256Manager.encrypt(emailInput);
+			return !memberRepository.existsByEmail(encryptedEmailInput);
 		}
 
 	}
+
 }
